DhcpHandler documentation and improvements.
The path where to store pickled configuration and generated config files
is configurable when constructing the instance.

If there's a pickled configuration it's automatically loaded at
construction time. If not, a new configuration based on defaults is dumped
to disk.

New methods are added: restart(), reload() and rollback() and DhcpHandler
is published in the services module and included in the pymin daemon
command tree.