Add help messages and command listing to dispatcher handlers.
Now the @handle decorator takes a mandatory help argument to specify a
short help string for that handler. A new functions get_help() is added to
the dispatcher module to get the help string from a handler.

A new base Handler class is provided in dispatcher module to automatically
have a 'help' and 'commands' commands available.